refactor

+++ Bangumi.Api/Common/Extensions.cs
@@ -13,11 +13,11 @@ namespace Bangumi.Api.Common
         /// </summary>
         /// <param name="dateTime"></param>
         /// <returns></returns>
-        public static long ToJsTick(this DateTime dateTime)
+        public static int ToJsTick(this DateTime dateTime)
         {
             // C# 时间戳为 1/10000000 秒，从0001年1月1日开始；Bangumi 使用的时间戳为秒，从1970年1月1日开始
             // 默认 DateTime 使用 UTC 时间
-            return (long)(dateTime.ToUniversalTime() - new DateTime(1970, 1, 1)).TotalSeconds;
+            return (int)(dateTime.ToUniversalTime() - new DateTime(1970, 1, 1)).TotalSeconds;
         }
 
         /// <summary>
@@ -25,7 +25,7 @@ namespace Bangumi.Api.Common
         /// </summary>
         /// <param name="dateTime"></param>
         /// <returns></returns>
-        public static DateTime ToDateTime(this long dateTime)
+        public static DateTime ToDateTime(this int dateTime)
         {
             // C# 时间戳为 1/10000000 秒，从0001年1月1日开始；Bangumi 使用的时间戳为秒，从1970年1月1日开始
             // 默认 DateTime 使用 UTC 时间
